requirements.txt
numpy
pandas
ipykernel
scikit-learn
matplotlib
xgboost
catboost
python-dotenv